requirements.txt
requests==2.31.0
pytrends==4.9.2
google-api-python-client==2.111.0
python-dotenv==1.0.0
